+++ cloudpassage/http_helper.py
@@ -6,6 +6,7 @@ import json
 import urlparse
 from cloudpassage.exceptions import CloudPassageValidation
 import cloudpassage.utility as utility
+from cloudpassage.retry import Retry
 import requests
 
 
@@ -82,6 +83,14 @@ class HttpHelper(object):
                                                       response.text)
             if success is True:
                 return response.json()
+
+        if response.status_code >= 500:
+            success, response, exception = Retry().get(url,
+                                                       headers,
+                                                       kwargs["params"])
+            if success is True:
+                return response.json()
+
         raise exception
 
     def get_paginated(self, endpoint, key, max_pages, **kwargs):
@@ -184,21 +193,27 @@ class HttpHelper(object):
                                  data=json.dumps(reqbody))
         success, exception = utility.parse_status(url, response.status_code,
                                                   response.text)
-        if success is False:
-            # If we get a 401, it could be an expired key.  We retry once.
-            if response.status_code == 401:
-                self.connection.authenticate_client()
-                headers = self.connection.build_header()
-                response = requests.post(url, headers=headers,
-                                         data=json.dumps(reqbody))
-                success, exception = utility.parse_status(url,
-                                                          response.status_code,
-                                                          response.text)
-                if success is True:
-                    return response.json()
-            raise exception
-        else:
+        if success is True:
             return response.json()
+
+        # If we get a 401, it could be an expired key.  We retry once.
+        if response.status_code == 401:
+            self.connection.authenticate_client()
+            headers = self.connection.build_header()
+            response = requests.post(url, headers=headers,
+                                     data=json.dumps(reqbody))
+            success, exception = utility.parse_status(url,
+                                                      response.status_code,
+                                                      response.text)
+            if success is True:
+                return response.json()
+
+        if response.status_code >= 500:
+            success, response, exception = Retry().post(url, headers, reqbody)
+            if success is True:
+                return response.json()
+
+        raise exception
 
     def put(self, endpoint, reqbody):
         """This method performs a PUT against Halo's API.
@@ -237,6 +252,14 @@ class HttpHelper(object):
                                                           response.text)
                 if success is True:
                     return response.json()
+
+            if response.status_code >= 500:
+                success, response, exception = Retry().put(url,
+                                                           headers,
+                                                           reqbody)
+                if success is True:
+                    return response.json()
+
             raise exception
         else:
             # Sometimes we don't get json back...
@@ -289,6 +312,12 @@ class HttpHelper(object):
                                                           response.text)
                 if success is True:
                     return response.json()
+
+            elif response.status_code >= 500:
+                success, response, exception = Retry().delete(url, headers)
+                if success is True:
+                    return response.json()
+
             raise exception
         else:
             # Sometimes we don't get json back...